@@ -13,14 +13,14 @@ $arrLang = $GLOBALS['TL_LANG']['tl_module'];
 /**
  * Fields
  */
-$arrLang['addSubmissionRelation'] = array('Einsendungen verknüpfen', 'Verknüpfen Sie Einsendungen mit anderen Contao-Entitäten wie z.B. Nachrichten.');
-$arrLang['submissionRelation'] = array('Typ der Verknüpfung', 'Wählen den Typ der Verknüpfung aus. Einsendungen werden anschließend mit den zugehörigen Entitäten verknüpft. (siehe $GLOBALS[\'SUBMISSION_RELATIONS\'])');
-$arrLang['limitSubmissionPeriod'] = array('Einreichungszeitraum eingrenzen', 'Grenzen Sie den Zeitraum ein in dem Einsendungen eingereicht werden können.');
-$arrLang['submissionStart'] = array('Einreichen ab', 'Einreichen von Einsendungen erst ab diesem Tag erlauben.');
-$arrLang['submissionStop'] = array('Einreichen bis', 'Einreichen von Einsendungen bis zu diesem Tag erlauben.');
+$arrLang['addSubmissionRelation'] = ['Einsendungen verknüpfen', 'Verknüpfen Sie Einsendungen mit anderen Contao-Entitäten wie z.B. Nachrichten.'];
+$arrLang['submissionRelation']    = ['Typ der Verknüpfung', 'Wählen den Typ der Verknüpfung aus. Einsendungen werden anschließend mit den zugehörigen Entitäten verknüpft. (siehe $GLOBALS[\'SUBMISSION_RELATIONS\'])'];
+$arrLang['limitSubmissionPeriod'] = ['Einreichungszeitraum eingrenzen', 'Grenzen Sie den Zeitraum ein in dem Einsendungen eingereicht werden können.'];
+$arrLang['submissionStart']       = ['Einreichen ab', 'Einreichen von Einsendungen erst ab diesem Tag erlauben.'];
+$arrLang['submissionStop']        = ['Einreichen bis', 'Einreichen von Einsendungen bis zu diesem Tag erlauben.'];
 
 /**
  * Legends
  */
 $arrLang['relation_legend'] = 'Verknüpfungs-Einstellungen';
-$arrLang['period_legend'] = 'Einsendungszeitraum-Einstellungen';
+$arrLang['period_legend']   = 'Einsendungszeitraum-Einstellungen';
