@@ -13,11 +13,11 @@ $arrLang = &$GLOBALS['TL_LANG']['tl_submission_relation_spread'];
 /**
  * Fields
  */
-$arrLang['addSubmissionRelation'] = array('Einsendungen verknüpfen', 'Stellen Sie eine Verknüpfung zu einem Einsendungen-Archiv her');
-$arrLang['submissionRelation'] = array('Einsendungs-Archiv', 'Wählen Sie ein Archiv aus, in das Einsendungen abgelegt werden.');
-$arrLang['limitSubmissionPeriod'] = array('Einreichungszeitraum eingrenzen', 'Grenzen Sie den Zeitraum ein in dem Einsendungen eingereicht werden können.');
-$arrLang['submissionStart'] = array('Einreichen ab', 'Einreichen von Einsendungen erst ab diesem Tag erlauben.');
-$arrLang['submissionStop'] = array('Einreichen bis', 'Einreichen von Einsendungen bis zu diesem Tag erlauben.');
+$arrLang['addSubmissionRelation'] = ['Einsendungen verknüpfen', 'Stellen Sie eine Verknüpfung zu einem Einsendungen-Archiv her'];
+$arrLang['submissionRelation']    = ['Einsendungs-Archiv', 'Wählen Sie ein Archiv aus, in das Einsendungen abgelegt werden.'];
+$arrLang['limitSubmissionPeriod'] = ['Einreichungszeitraum eingrenzen', 'Grenzen Sie den Zeitraum ein in dem Einsendungen eingereicht werden können.'];
+$arrLang['submissionStart']       = ['Einreichen ab', 'Einreichen von Einsendungen erst ab diesem Tag erlauben.'];
+$arrLang['submissionStop']        = ['Einreichen bis', 'Einreichen von Einsendungen bis zu diesem Tag erlauben.'];
 
 /**
  * Legends
