@@ -13,5 +13,5 @@ $arrLang = &$GLOBALS['TL_LANG']['tl_submission'];
 /**
  * Fields
  */
-$arrLang['news'] = array('News', 'Select a news from a relation for submission.');
-$arrLang['news'] = array('Event', 'Select an event from a relation for submission.');
+$arrLang['news'] = ['News', 'Select a news from a relation for submission.'];
+$arrLang['news'] = ['Event', 'Select an event from a relation for submission.'];
