@@ -13,11 +13,11 @@ $arrLang = &$GLOBALS['TL_LANG']['tl_submission_relation_spread'];
 /**
  * Fields
  */
-$arrLang['addSubmissionRelation'] = array('Link submissions', 'Make a link to an submissions archive.');
-$arrLang['submissionRelation'] = array('Submission archive', 'Select an archive where submissions will be stored in.');
-$arrLang['limitSubmissionPeriod'] = array('Limit submit period', 'Limit the submit submissions period.');
-$arrLang['submissionStart'] = array('Submit from', 'Do not allow submit of submissions before this day.');
-$arrLang['submissionStop'] = array('Submit until', 'Do not allow submit of submissions after this day.');
+$arrLang['addSubmissionRelation'] = ['Link submissions', 'Make a link to an submissions archive.'];
+$arrLang['submissionRelation']    = ['Submission archive', 'Select an archive where submissions will be stored in.'];
+$arrLang['limitSubmissionPeriod'] = ['Limit submit period', 'Limit the submit submissions period.'];
+$arrLang['submissionStart']       = ['Submit from', 'Do not allow submit of submissions before this day.'];
+$arrLang['submissionStop']        = ['Submit until', 'Do not allow submit of submissions after this day.'];
 
 
 /**
