@@ -11,5 +11,5 @@
 /**
  * Front end modules
  */
-$GLOBALS['TL_LANG']['MSC']['submission_news_inactive'] = 'Currently no new submission for "%s" possible.';
+$GLOBALS['TL_LANG']['MSC']['submission_news_inactive']  = 'Currently no new submission for "%s" possible.';
 $GLOBALS['TL_LANG']['MSC']['submission_event_inactive'] = 'Currently no participation in the event "%s" possible.';
